make mops data

+++ mops/mops_totals_over_years.py
@@ -28,7 +28,7 @@ df_fire = df[df.iloc[:, 9].isin(ministry_codes[4])]
 df_list = [df_ministry, df_witness, df_police, df_prison, df_fire]
 
 budget_sums = [(year, budget_type, data[(data.iloc[:, 0] == year) &
-                                        (data.iloc[:, 21] == budget_type)].iloc[:, 22].sum())
+                                        (data.iloc[:, 21] == budget_type)].iloc[:, 22].astype(int).sum())
                for data in df_list
                for year in available_years
                for budget_type in types]
